Transliteration_service/requirements.txt
ai4bharat-transliteration
gradio
